feat(instance): add server action command
