+++ cmd/scw/testdata/test-all-usage-instance-server-action-usage.golden
@@ -1,0 +1,38 @@
+🎲🎲🎲 EXIT CODE: 0 🎲🎲🎲
+🟥🟥🟥 STDERR️️ 🟥🟥🟥️
+Perform an action on a server
+
+USAGE:
+  scw instance server action <server-id ...> [arg=value ...]
+
+EXAMPLES:
+  Start a server in the default zone with a given id
+    scw instance server action 11111111-1111-1111-1111-111111111111 action=start
+
+ARGS:
+  action            The action to perform
+  server-id         ID of the server affected by the action.
+  [zone=fr-par-1]   Zone to target. If none is passed will use default zone from the config
+
+FLAGS:
+  -h, --help   help for action
+  -w, --wait   wait until the server is ready
+
+GLOBAL FLAGS:
+  -c, --config string    The path to the config file
+  -D, --debug            Enable debug mode
+  -o, --output string    Output format: json or human, see 'scw help output' for more info (default "human")
+  -p, --profile string   The config profile to use
+
+SEE ALSO:
+  # List available actions for a server
+  scw instance server list-actions
+
+  # Perform reboot action
+  scw instance server reboot
+
+  # Perform start action
+  scw instance server start
+
+  # Perform stop action
+  scw instance server stop
