+++ internal/namespaces/instance/v1/testdata/test-server-action-manual-poweron.golden
@@ -1,0 +1,186 @@
+🎲🎲🎲 EXIT CODE: 0 🎲🎲🎲
+🟩🟩🟩 STDOUT️ 🟩🟩🟩️
+ID                            fbbdbbf9-8130-4497-9c75-ceb34c61de08
+Name                          cli-srv-ecstatic-bohr
+Organization                  ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b
+Project                       ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b
+AllowedActions.0              poweroff
+AllowedActions.1              terminate
+AllowedActions.2              reboot
+AllowedActions.3              stop_in_place
+AllowedActions.4              backup
+CommercialType                DEV1-S
+CreationDate                  few seconds ago
+DynamicIPRequired             true
+RoutedIPEnabled               false
+EnableIPv6                    false
+Hostname                      cli-srv-ecstatic-bohr
+Image.ID                      08892932-3f2a-4f8a-a9ab-4a5b52630c57
+Image.Name                    Ubuntu 22.04 Jammy Jellyfish
+Image.Arch                    x86_64
+Image.CreationDate            few seconds ago
+Image.ModificationDate        few seconds ago
+Image.ExtraVolumes            0
+Image.FromServer              -
+Image.Organization            51b656e3-4865-41e8-adbc-0c45bdd780db
+Image.Public                  true
+Image.RootVolume              9549ab62-3517-4ecd-bd23-432d797f57bd
+Image.State                   available
+Image.Project                 51b656e3-4865-41e8-adbc-0c45bdd780db
+Image.Zone                    fr-par-1
+Protected                     false
+PrivateIP                     10.73.104.133
+PublicIP.ID                   e6b4161c-85e3-4d3b-bdcc-f881afdf7c46
+PublicIP.Address              163.172.133.158
+PublicIP.Netmask              32
+PublicIP.Family               inet
+PublicIP.Dynamic              false
+PublicIP.ProvisioningMode     dhcp
+PublicIPs.0.ID                e6b4161c-85e3-4d3b-bdcc-f881afdf7c46
+PublicIPs.0.Address           163.172.133.158
+PublicIPs.0.Netmask           32
+PublicIPs.0.Family            inet
+PublicIPs.0.Dynamic           false
+PublicIPs.0.ProvisioningMode  dhcp
+MacAddress                    de:00:00:1e:f5:b5
+ModificationDate              few seconds ago
+State                         running
+Location                      fr-par-1
+Bootscript                    x86_64 mainline 4.4.230 rev1
+BootType                      local
+Volumes.0.ID                  de8bf61b-3cee-4378-8ddf-b8d6217b439d
+Volumes.0.Name                Ubuntu 22.04 Jammy Jellyfish
+Volumes.0.ExportURI           -
+Volumes.0.Organization        ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b
+Volumes.0.Server.ID           fbbdbbf9-8130-4497-9c75-ceb34c61de08
+Volumes.0.Server.Name         cli-srv-ecstatic-bohr
+Volumes.0.Size                20 GB
+Volumes.0.VolumeType          l_ssd
+Volumes.0.CreationDate        few seconds ago
+Volumes.0.ModificationDate    few seconds ago
+Volumes.0.State               available
+Volumes.0.Project             ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b
+Volumes.0.Boot                false
+Volumes.0.Zone                fr-par-1
+SecurityGroup.ID              0fe819c3-274d-472a-b3f5-ddb258d2d8bb
+SecurityGroup.Name            Default security group
+StateDetail                   booting kernel
+Arch                          x86_64
+Zone                          fr-par-1
+🟩🟩🟩 JSON STDOUT 🟩🟩🟩
+{
+  "id": "fbbdbbf9-8130-4497-9c75-ceb34c61de08",
+  "name": "cli-srv-ecstatic-bohr",
+  "organization": "ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b",
+  "project": "ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b",
+  "allowed_actions": [
+    "poweroff",
+    "terminate",
+    "reboot",
+    "stop_in_place",
+    "backup"
+  ],
+  "tags": [],
+  "commercial_type": "DEV1-S",
+  "creation_date": "1970-01-01T00:00:00.0Z",
+  "dynamic_ip_required": true,
+  "routed_ip_enabled": false,
+  "enable_ipv6": false,
+  "hostname": "cli-srv-ecstatic-bohr",
+  "image": {
+    "id": "08892932-3f2a-4f8a-a9ab-4a5b52630c57",
+    "name": "Ubuntu 22.04 Jammy Jellyfish",
+    "arch": "x86_64",
+    "creation_date": "1970-01-01T00:00:00.0Z",
+    "modification_date": "1970-01-01T00:00:00.0Z",
+    "default_bootscript": null,
+    "extra_volumes": {},
+    "from_server": "",
+    "organization": "51b656e3-4865-41e8-adbc-0c45bdd780db",
+    "public": true,
+    "root_volume": {
+      "id": "9549ab62-3517-4ecd-bd23-432d797f57bd",
+      "name": "Ubuntu 22.04 Jammy Jellyfish",
+      "size": 10000000000,
+      "volume_type": "unified"
+    },
+    "state": "available",
+    "project": "51b656e3-4865-41e8-adbc-0c45bdd780db",
+    "tags": [],
+    "zone": "fr-par-1"
+  },
+  "protected": false,
+  "private_ip": "10.73.104.133",
+  "public_ip": {
+    "id": "e6b4161c-85e3-4d3b-bdcc-f881afdf7c46",
+    "address": "163.172.133.158",
+    "netmask": "32",
+    "family": "inet",
+    "provisioning_mode": "dhcp"
+  },
+  "public_ips": [
+    {
+      "id": "e6b4161c-85e3-4d3b-bdcc-f881afdf7c46",
+      "address": "163.172.133.158",
+      "netmask": "32",
+      "family": "inet",
+      "provisioning_mode": "dhcp"
+    }
+  ],
+  "mac_address": "de:00:00:1e:f5:b5",
+  "modification_date": "1970-01-01T00:00:00.0Z",
+  "state": "running",
+  "location": {
+    "cluster_id": "34",
+    "hypervisor_id": "701",
+    "node_id": "67",
+    "platform_id": "14",
+    "zone_id": "par1"
+  },
+  "ipv6": null,
+  "bootscript": {
+    "bootcmdargs": "LINUX_COMMON scaleway boot=local nbd.max_part=16",
+    "default": true,
+    "dtb": "",
+    "id": "fdfe150f-a870-4ce4-b432-9f56b5b995c1",
+    "initrd": "http://10.194.3.9/initrd/initrd-Linux-x86_64-v3.14.6.gz",
+    "kernel": "http://10.194.3.9/kernel/x86_64-mainline-lts-4.4-4.4.230-rev1/vmlinuz-4.4.230",
+    "organization": "11111111-1111-4111-8111-111111111111",
+    "project": "11111111-1111-4111-8111-111111111111",
+    "public": true,
+    "title": "x86_64 mainline 4.4.230 rev1",
+    "arch": "x86_64",
+    "zone": "fr-par-1"
+  },
+  "boot_type": "local",
+  "volumes": {
+    "0": {
+      "id": "de8bf61b-3cee-4378-8ddf-b8d6217b439d",
+      "name": "Ubuntu 22.04 Jammy Jellyfish",
+      "export_uri": "",
+      "organization": "ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b",
+      "server": {
+        "id": "fbbdbbf9-8130-4497-9c75-ceb34c61de08",
+        "name": "cli-srv-ecstatic-bohr"
+      },
+      "size": 20000000000,
+      "volume_type": "l_ssd",
+      "creation_date": "1970-01-01T00:00:00.0Z",
+      "modification_date": "1970-01-01T00:00:00.0Z",
+      "state": "available",
+      "project": "ee7bd9e1-9cbd-4724-b2f4-19e50f3cf38b",
+      "boot": false,
+      "zone": "fr-par-1"
+    }
+  },
+  "security_group": {
+    "id": "0fe819c3-274d-472a-b3f5-ddb258d2d8bb",
+    "name": "Default security group"
+  },
+  "maintenances": [],
+  "state_detail": "booting kernel",
+  "arch": "x86_64",
+  "placement_group": null,
+  "private_nics": [],
+  "zone": "fr-par-1"
+}
